requirements.txt
azure-core==1.30.2
azure-storage-blob==12.22.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Cython==3.0.11
fonttools==4.53.1
idna==3.7
isodate==0.6.1
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
statsmodels @ git+https://github.com/statsmodels/statsmodels.git@eef6ad3d5f18862877c691a2fa25a57df70c1aa5
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
